websites/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
jinja2==3.1.2
pillow==10.1.0
python-jose[cryptography]==3.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.12.1
httpx==0.25.2
aiofiles==23.2.1
